+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+### v4.0.1
+Add custom variable template to Tag Manager
+- Click data-attribute
+
 ### v4.0.0
 Add tags to Tag Manager
 - Axeptio